(Try to) simplify CLOEXEC setting in DefaultStream
Commit 08cc853b86 introduced code to fix compilation on non-Linux. Part
of the problem was that the recvmsg() flag MSG_CMSG_CLOEXEC does not
exist everywhere. The resulting code felt a bit spaghetti-sh to me, so
this commit tries to clean that up.

This commit adds a new recvmsg helper module that abstracts away these
differences. There are two versions of this module: One for systems with
MSG_CMSG_CLOEXEC and one for without.

No functional changes intended from this commit, except that a possible
FD leak if fcntl() fails is fixed.

Signed-off-by: Uli Schlachter <psychon@znc.in>