Ignore examples for code coverage
Tracking code coverage for examples is not interesting. No one should
write tests for them.

Also, if we were to include the execution of these examples (easily
possible!), this would also consider all code in x11rb and
x11rb-protocol that is used by these examples to be covered. I consider
the examples "too light a test" to really count the code they use as
properly tested.

Signed-off-by: Uli Schlachter <psychon@znc.in>