Try to stabilise code coverage reports
We got some flaky code coverage results. The resource database code
sometimes takes different paths. Apparently, $XENVIRONMENT is sometimes
set and sometimes is not.

This commit tries to get rid of the noise in coverage reports by simply
always setting $XENVIRONMENT to a file that does not exist.

Time will tell if this works properly. No idea how to test this.

Signed-off-by: Uli Schlachter <psychon@znc.in>