Simplify/prettify tutorial example
The recent commit 6db1a8a72 made the code in the print_modifiers()
example a bit uglier. That code position already had a "FIXME: This
should be made nicer" comment.

Turns out that we already made this nicer a while back and the Debug
impl for bitmasks decodes the bitmask and prints the individual members.
That is exactly what print_modifiers() needs.

Example of output before this commit:

  Modifier mask: ["Shift", "Ctrl", "Button1"]

Example of output after this commit:

  Modifier mask: Shift | Control | Button1

The commit uses the alternate form of the debug output (modifier #)
since that one writes e.g. "Shift" instead of "SHIFT" and that is closer
to the previous output.

Signed-off-by: Uli Schlachter <psychon@znc.in>