Add functions to bitmask_binop
The bitmask_binop macro is invoked for enumerations that "look like" bit
masks. This commit adds a contains() associated function that checks if
"this" mask has all bits set for a given mask.

Additionally, this adds implementations for BitOrAssign and BitAndAssign
for such an enum with itself. E.g. previously EventMask::NO_EVENT |
EventMask::KEY_PRESS would not compile.

Finally, this adds some superficial tests for the bitmask_binop
functions.

Signed-off-by: Uli Schlachter <psychon@znc.in>