xinput::EventForSend: Do not hide errors
The only reason that <eventdef> exists in the XML is for xinput. The
generated code for this contained calls to unwrap(). This commit changes
that to pass possible ParseError instances to the caller.

This PR related to #178 ("Get rid of unwrap() in the generated code")
and #49 ("Do something about the results of 'git grep FIXME'").