Improve names of multi-field bitcase structs
When a <case> or <bitcase> of a <switch> contains more than one field,
we have to generate a struct that holds these fields. This struct needs
a name. Up to now, we used something ugly like "bitcase8" to generate
this name, leading to such beautiful struct names as
SelectEventsAuxBitcase8.

This commit changes the code to consider an <enumref> in the case. This
has to refer to some enum to represent that value that indicates that
this case is present. The variant of this enum has a name. This name is
now used to generate the name. SelectEventsAuxBitcase8 for example is
now called SelectEventsAuxBellNotify instead.

Fixes: https://github.com/psychon/x11rb/issues/156
Signed-off-by: Uli Schlachter <psychon@znc.in>