Get rid of unwrap() for array parsing in generated code
The generated code for parsing an array first got a slice of the
expected length and then used the TryFrom trait to get an array. For
example:

  let (data1, remaining) = crate::x11_utils::parse_u8_list(remaining, 24)?;
  let data1 = <[u8; 24]>::try_from(data1).unwrap();

This commit introduces a helper function that produces an array directly. The
above becomes:

  let (data1, remaining) = crate::x11_utils::parse_u8_array::<24>(remaining)?;

The goal here is to remove some unwrap()s in the generated code, but
this also helps shortening the code.

There are two new helper function: parse_u8_array parses an array while
parse_u8_array_ref returns a reference to an array. Both cases are
needed in the generated code.

Signed-off-by: Uli Schlachter <psychon@znc.in>