Fix sequence number wrap around detection
On the wire, packets only have 16 bits for a sequence number. Thus, this
can overflow / wrap around. To ensure that we can still reconstruct
sequence numbers, there needs to be enough requests with a reply so that
we can know when the overflow happened.

The previous code allows 2^16 such requests. This was not enough: If
both replies have a 16 bit sequence number of "1", we do not know if the
first request had multiple replies or if the second reply is for the
second request.

Thus, this commit changes the code to require a request with a reply
after 2^16 - 1 requests.

Also, a test is added that actually "fakes" replies and tests that the
correct sequence number was reconstructed for the second request with a
16 bit sequence number of "1". This new test obviously fails without the
fix. Adding extra "println!" calls shows that the sequence number of the
second reply is incorrectly reconstructed to be 1 instead of 0x10001
(without the fix).

Signed-off-by: Uli Schlachter <psychon@znc.in>