Fix dl-libxcb feature on OpenBSD and NetBSD
Turns out that libtool uses a different logic for turning its
"-version-info" into a file name on OpenBSD and NetBSD than on Linux,
Darwin, and FreeBSD. On the affected BSDs, this just takes "current",
which is the first argument to "-version-info". Everywhere else, the
logic is "current - age".

Since libxcb uses "-version-info 2:0:1", this means that we end up with
libxcb.so.2 on OpenBSD and NetBSD instead of the libxcb.so.1 that the
code expects. Thus, we failed to load libxcb.

Fix this by correcting the library name on OpenBSD and NetBSD.

Additionally, the folder is apparently not in the default search path,
so we need an absolute path. No idea whether this is also needed on
OpenBSD. I will wait for someone to complain/test.

Fixes: https://github.com/psychon/x11rb/issues/785
Signed-off-by: Uli Schlachter <psychon@znc.in>