Split out shared state from async RustConnection
This commit moves the code for the SharedState and PacketReader in
x11rb-async from rust_connection/mod.rs to a new submodule in an extra
file. The goal here is to make that file smaller and hopefully less
complex.

While here, I introduced a wrapper around inner.lock().unwrap() to lock
the ProtoConnection and another wrapper around the loop for waiting for
a certain incoming pattern called wait_for_incoming().

Signed-off-by: Uli Schlachter <psychon@znc.in>