Split out the write buffer from async RustConnection
This moves the write buffer implementation in x11rb-async from
rust_connection/mod.rs to a new submodule in an extra file. The goal
here is to make that file smaller and hopefully less complex.

No changed in behaviour are intended. However, WriteBuffer now
implements Default and WriteBufferGuard implements Deref and DerefMut.
This allows to call flush() and write_all_vectored() on
WriteBufferInner.

Signed-off-by: Uli Schlachter <psychon@znc.in>