Simplify locking in async check_for_raw_error()
THis commit (hopefully) simplifies the locking in
RustConnection::check_for_raw_error().

Previously, the inner connection state was first checked and updated (is
a sync needed?) and only afterwards was the write buffer locked for
sending this sync. Now, the write buffer is immediately locked.

This allows getting rid of some "did we already lock the write buffer or
do we need to lock it now" code before flushing the buffer. Hopefully
this also makes the code more correct.

This code now acquires locks in the same order as send_request(): First,
lock the write buffer and only afterwards is the inner connection is
locked. There was no deadlock in this code before (since the lock from
lock_connection() was already dropped by the time the write buffer was
locked), but it still feels a bit more future-safe to have the lock
order in both functions be the same.

Signed-off-by: Uli Schlachter <psychon@znc.in>