x11rb-async: Break when drive future is dropped
This commit changes the implementation of timeouts in the xclock_utc
example of x11rb-async. Previously, it simply called exit(0) on timeout.
This commit changes things to cancel the drive future on a timeout.

From the point of view of the driving future, this simply means that the
future is suddenly dropped and no longer polled. This means that
x11rb-async stops reading input from the X11 connection, but sending
would just work.

For the xclock_utc example, this means that the clock would still tick,
but it would no longer redraw itself in response to expose events and
would not react to resizes. This is quite confusing behaviour.

Now we come to what I actually want to achieve (the above is just
preparation to show how confusing this case can be and to have a test
case):

A new object is added and saved in the future used by drive(). When this
object is dropped, it marks the connection as broken by setting an
AtomicBool and wakes up all futures waiting for input. The function
wait_for_incoming() is updated to check this flag. Its return type gets
an extra Result and a dropped driving future is turned into an I/O
error.

Example output of the xclock_utc example on a timeout:

Cancelling drive task due to $X11RB_EXAMPLE_TIMEOUT
Fatal Error: Driving future was dropped

Signed-off-by: Uli Schlachter <psychon@znc.in>