Add a test for "immediately dropping the driver"
This commit adds a test case that checks that immediately dropping the
driver future without polling it once breaks the connection.

Why does this get a special test? Because the usual CallOnDrop
implementation that constructs some object inside of an "async fn" does
not work if that future is not even polled once. This special case is
surprising enough that it deserves a test.

Signed-off-by: Uli Schlachter <psychon@znc.in>