Fix a logic error in WriteBuffer
The write buffer should continue writing as long as there are more bytes from
the input *or* a partial write caused data to be split out from the input.
Instead of the "or", the code contained an "and". Since "partial" is initially
empty, this caused the code to just not write anything at all.

Fixes: https://github.com/psychon/x11rb/issues/810
Signed-off-by: Uli Schlachter <psychon@znc.in>