Do not silently drop FDs if FD passing fails
This commit adds a new test. This test sends FDs over a stream, but the
stream does not actually send and consume the FDs.

Previously, this condition was silently ignored. This commit makes the
WriteBuffer retry the write, which results in an IOError with
ErrorKind::WriteZero. This might not be the best error message, but
matches what x11rb does.

Signed-off-by: Uli Schlachter <psychon@znc.in>