Improve error when not all FDs can be sent
An implementation of Stream is supposed to send all the FDs that we give
to it. When it would not send some FDs, we had a loop that retries the
sending, but with zero bytes of data. This necessarily resulted in a
WriteZero error, which is not nice.

This commit changes the code instead to produce an IOError with a better
error message.

Signed-off-by: Uli Schlachter <psychon@znc.in>