Fix use-after-move without "tracing"
This fixes the following error:

error: format argument must be a string literal
   --> x11rb/src/rust_connection/mod.rs:655:27
    |
655 |             crate::trace!({ sequence }, "Polling for reply");
    |                           ^^^^^^^^^^^^

Previously, I was getting a warning about an unused variable, which is
why I created the "let _ ="-hack. However, this moves variables.

In this commit, the format string is now passed to format_args!()
instead to have the format arg properly handled. This means that we need
to be able to detect the format strings and ignore other fields. For
this, the extra "{}" is used for detection.

Signed-off-by: Uli Schlachter <psychon@znc.in>