Reduce logging "costs" if logging is disabled
The code in send_request() always figured out some information about the
request being sent and then logged it. Even if the log level was
disabled.

This commit moves the "figure things out" into a Display impl. That way,
this information is not actually being collected if logging is disabled.

Signed-off-by: Uli Schlachter <psychon@znc.in>