Update to actions/checkout@v3
From lots of CI build logs:

> Node.js 12 actions are deprecated. Please update the following actions
> to use Node.js 16: actions/checkout@v2, actions-rs/toolchain@v1. For
> more information see:
> https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/.

The readme of actions/checkout says under "What's new" that they updated to
node16. So, that should be a step towards fixing the above error.

We still use lots of other actions that were not updated in a long time, so
there is still some work ahead.

Signed-off-by: Uli Schlachter <psychon@znc.in>