Deduplicate request name decoding
Recently, I introduced a new function get_request_name() to get a human
readable name of a request based on a request's major and minor opcode.
This function is used in the tracing code in x11rb.

However, we already had a similar function request_name(). This function
was not public and was used by X11Error::try_parse() so that we provide
a human-readable representation of the request name that caused an X11
error.

Since an X11Error has an extra field for the name of the extension,
request_name() only returned the actual name of the request.
get_request_name() instead also included the name of the extension in
its result.

This commit deduplicates this. Instead of having two functions that do
almost the same thing, there is a new function
get_request_name_internal() that returns an internal enum. The other two
functions are just short wrappers around this internal function.

One complication comes from the fact that X11Error contains the name of
the extension in an extra field. I didn't want to introduce two calls to
ExtInfoProvider::get_from_major_opcode() that both look up the same
information. Previously, a call to this function was in
X11Error::try_parse() and also in get_extension_version(). I refactored
some code so that the new get_request_name_internal() also returns this
information so that we can continue to do only a single lookup.

Signed-off-by: Uli Schlachter <psychon@znc.in>