x11rb::properties: Indicate missing properties
Right now, the from_reply() method of WmClass, WmSizeHints, and WmHints
returns Err(ParseError::InvalidValue) when the GetPropertyReply does not
have the expected type and format. However, one way in which this can
happen is when a property is queried that is not present. In this case,
an InvalidValue error is not really nice or helpful.

This is especially unusable if using the reply() method on the
corresponding cookie returns this error, since the caller has no way to
check whether this is a "real" parse error or just a missing property.

This commit adds an extra Option to the result of from_reply(). This
Option is None if the GetPropertyReply indicates that the property does
not exist.

The cookie object's reply() method is changed accordingly and also wraps
its return type into an extra Option().

The reply_unchecked() method is not changed. It already turns all
parsing errors into None. Thus, "property not present" can just be
"swallowed" into this None as well.

Signed-off-by: Uli Schlachter <psychon@znc.in>