Enable all features on docs.rs
Currently we have to ensure that the docs on docs.rs are actually
complete and no optional stuff is left out. Recently I saw an option to
just enable all-features and I really don't know why anyone wouldn't
want this. And according to [0] this really exists!

So, just enable all features when building on docs.rs.

[0]: https://docs.rs/about/metadata

Signed-off-by: Uli Schlachter <psychon@znc.in>