Remove PiecewiseBuf and use array instead
Before this commit, serialising a request returns a
BufWithFds<PiecewiseBuf<'some_lifetime>>, which expands to
(Vec<Cow<'some_lifetime, [u8]>>, Vec<RawFdContainer>). That's a lot of
Vecs and most of them are not necessary since we statically know how
many "pieces" a request needs.

Side note: Why not just use a single Vec<u8>? There are some situations
where we can avoid copying some data. Think e.g. about PutImage.

This commit removes PiecewiseBuf (which is a Vec) and instead makes the
code generator use an array. This is possible since we statically know
the number of pieces for request sending.

The benefit is that this avoids a temporary allocation for the Vec.

Signed-off-by: Uli Schlachter <psychon@znc.in>