Use array for IoSlice construction
Before this commit, the request functions in x11rb and x11rb-async
contained code like this:

    let (bytes, fds) = request0.serialize(major_opcode(conn).await?);
    let slices = bytes.iter().map(|b| IoSlice::new(b)).collect::<Vec<_>>();
    conn.send_request_without_reply(&slices, fds).await

Here, "bytes" is an array (thanks to the previous commit) and each of its
entries is wrapped in an IoSlice. The result is saved in a Vec.

This commit replaces the above with the following code:

    let (bytes, fds) = request0.serialize(major_opcode(conn).await?);
    let slices = [IoSlice::new(&bytes[0])];
    assert_eq!(slices.len(), bytes.len());
    conn.send_request_without_reply(&slices, fds).await

Instead of a temporary allocation with a Vec, another array is created.
I found no generic way to create this array, so instead the code
generator "knows" what the length of the array must be and repeats each
case as necessary.

Just to double-check that we got this right, an assertion is added to
check that "bytes" and "slices" has the same length. Since both are
arrays, their length is a constant and this assertion can hopefully be
optimised away by the compiler.

Signed-off-by: Uli Schlachter <psychon@znc.in>