Add new parse_display_with_file_exists_callback() func
The logic for parsing a $DISPLAY value now depends on whether some given
file exists or not. To make this better testable and usable in no_std
mode, a new function parse_display_with_file_exists_callback() is added.
This works just like parse_display(), but has an extra argument that is
the callback function to use for checking whether some given string
refers to an existing file.

The existing parse_display() now is only available when "std" is
enabled. It uses the new function and provides a callback based on
std::path::Path::exists().

Signed-off-by: Uli Schlachter <psychon@znc.in>