Add GenerateRandom trait to x11rb-protocol
We want to have randomised round-trip testing. An instance of a struct
is generated randomly, then serialised, parsed and the result should be
equivalent to the original structure. Existing crates for this seem not
to support our use case. See [0] for details.

Thus, this commit starts work on our own implementation of this. It
begins by adding a GenerateRandom to x11rb-protocol that allows to
create a random instance of a type based on a fastrand random number
generator.

This commit also implements this trait for primitive types. Numbers are
limited to at most 4 since they could be sizes and we have nesting. For
example, if a list contains another list which contains another list,
with a limit of 4 we can already end up with 4*4*4 = 64 items. To avoid
generating huge instances, this low limit is used.

[0]: https://github.com/psychon/x11rb/issues/453

Signed-off-by: Uli Schlachter <psychon@znc.in>