Improve error messages for $DISPLAY parsing failures
So far, when we fail to parse $DISPLAY, the error message "Display
parsing error" would be generated. That's not saying too much helpful
things, besides "could you please tell me what value $DISPLAY is set
to?".

This commit changes the code to instead have an enumeration of error
cases. When parsing actually fails, the value that was attempted to be
parsed is now also included in the error message.

This would have made [0] much more obvious.

[0]: https://github.com/quininer/x11-clipboard/issues/38

Signed-off-by: Uli Schlachter <psychon@znc.in>