Use <[u8]>::escape_ascii()
Rust 1.60 added the escape_ascii() function on slices of u8. This
produces an iterator with items u8 that provides escapement in the style
of std::ascii::escape_default(). The returned iterator implements
Display, thus we can just pass it directly to write!() to get the
resulting escaped string.

The above allows to simplify some hand-written version doing the same.

Signed-off-by: Uli Schlachter <psychon@znc.in>