Remove enum from DefaultStream
DefaultStream only contains a DefaultStreamInner. So far this was an
enum that was used to remember what kind of stream we had. However, a FD
is a FD. We do not actually care about the kind of stream. Thus, this
commit replaces the enum DefaultStreamInner with a type alias to
RawFdContainer.

RawFdContainer only exists on Unix. On other systems, TcpStream is the
only supported transport. Thus, in that case we use a TcpStream instead.

Signed-off-by: Uli Schlachter <psychon@znc.in>