Avoid unwrap() in properties.rs
This code has a constant with value 18. So far, this constant was kept
as a u32. There is one place that needs this value as a usize. This
conversion was done with try_into().unwrap().

This commit removes that unwrap() by changing the type of the constant
to u16, which can be converted to usize via into().

Signed-off-by: Uli Schlachter <psychon@znc.in>