Implement RequestConnection for some types
This commit implements RequestConnection for &C, &mut C, Box<C>, Arc<C>,
Rc<C>, and Cow<'_, C>, where C: RequestConnection. All of these types
dereference to C. Thus, this can be implemented with a common macro.

The implementation for Cow looks different than the other ones since it
also requires C: ToOwned.

To make the implementation actually work, all cookie types get a new
replace_connection() helper function. This function replaces the
reference to a connection that is contained in the cookie. Without this
helper function, there is a type mismatch and the impl for
RequestConnection does not work.

Signed-off-by: Uli Schlachter <psychon@znc.in>