Wrapper types: Do not require references
This commit changes e.g. xproto::WindowWrapper to contain "C" instead of
"&C", i.e. a RequestConnection directly instead of just a reference to
one.

Since the previous commit implemented RequestConnection for references
to RequestConnection, this should not actually break anything (well, of
course user code breaks since the struct no longer contains a lifetime).

Since the previous commit also implemented RequestConnection for e.g.
Arc, this allows to use an Arc in a reference connection. So far, the
only way to construct such an instance is via
WindowWrapper::for_window(). The convenience function that creates a
window and returns a wrapper has Cookie involved in it, which makes it
more complicated.

Thus, this commit does not fully fix the following issue, but only
"helps with it".

Helps-with: https://github.com/psychon/x11rb/issues/870
Signed-off-by: Uli Schlachter <psychon@znc.in>