Wrapper types: Allow non-reference calls
This commit makes it possible to call e.g.
WindowWrapper::create_window() with a first argument of type
Arc<RustConnection>. Previously, the first argument always needed to be
a reference.

Previously, the implementation of this method forwarded the call to the
function "*_and_get_cookie". However, a Cookie always contains a
reference to the original connection and thus always needs a reference.
Since we want to get rid of this reference, the implementation is now
copied.

Fixes: https://github.com/psychon/x11rb/issues/870
Signed-off-by: Uli Schlachter <psychon@znc.in>