Fix unused_qualifications warnings
Commit 8c7697d17c9 introduced a "use" for BorrowedFd, which caused some
pre-existing code to now generate a unused_qualifications warning. This
was fixed with #[allow].

However, I think that just removing this qualification is the nicer fix.

If Windows wasn't a thing, this would be all that there is to it. But
this core also uses rustix::fd::BorrowedFd on Windows. So what is that
type actually?

On Windows, BorrowedFd is a re-export of BorrowedSocket from std with a
different name. On Unix, this really is BorrowedFd from std.

Thus, to make this work, we have to import BorrowedFd through the rustix
re-export instead of directly importing it from std.

Signed-off-by: Uli Schlachter <psychon@znc.in>