Really fix dl-libxcb feature on OpenBSD
In commit 53de935c I fixed the dl-libxcb feature not finding libxcb.so
on NetBSD and just assumed that this also fixes things on OpenBSD since
I found some similarities. However, it turns out that OpenBSD ignores
libtool's SONAME handling and just rolls its own.

However, we do not really have to care much. Here, loading "libxcb.so"
does the right thing. Not because there is actually a file with this
name, but since the dynamic linker does some magic when it gets a
request to load a library without providing version information. Thus,
this commit changes the code to just load "libxcb.so" on OpenBSD.

With that, I got thinking. On untested systems, we do not really know
which SONAME to use. And we now have evidence that somehow just letting
the OS do some magic is the best option. Thus, this commit changes the
code to load "libxcb.so" on all systems except for Linux (where this
would e.g. require libxcb1-dev to be installed on Debian) and NetBSD
(where we have a working version that I do not want to break).

Fixes: https://github.com/psychon/x11rb/issues/895
Signed-off-by: Uli Schlachter <psychon@znc.in>