Silence some new clippy lints
Clippy 1.75 started producing warnings like the following. Fixing this
properly by using From is not so simple since we would have to figure
out the involved types correctly. Instead, let's just silence this
clippy lint and hope that the compiler optimises away the unreachable
panic in the call to unwrap().

warning: use of a fallible conversion when an infallible one could be used
       --> x11rb-protocol/src/protocol/xkb.rs:12503:49
        |
  12503 |             assert_eq!(radio_group_names.len(), usize::try_from(n_radio_groups).unwrap(), "`radio_group_names` has an incorrect length");
        |                                                 ^^^^^^^^^^^^^^^ help: use: `From::from`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

Signed-off-by: Uli Schlachter <psychon@znc.in>