Remove extra-traits from x11rb-protocol default
This feature enables lots of extra trait implementations that are not
helping compile times much. Thus, it makes sense to not enable this by
default, but instead have everyone who needs this explicitly enable it.

To make this actually work, some tests needed to be disabled that depend
on these extra traits.

Signed-off-by: Uli Schlachter <psychon@znc.in>