Automate updates of doc/generated_code.md
This commit introduces a new crate to the workspace. This new code
parses the generated code and automatically updates
doc/generated_code.md based on the result.

This code is a mess of heuristics and regexes, but hopefully good enough
to save some time in the long run. We will see...

Signed-off-by: Uli Schlachter <psychon@znc.in>