Remove unnecessary API from StreamAdaptor
A StreamAdaptor is only constructed internally in
RustConnection::connect() and cannot be queried by the user. Thus, we
don't need much public API for it.

This commit removes everything that we do not use internally.

If people want to call RustConnection::connect_to_stream() with a
StreamAdaptor, that is still possible, because I left new() public.

Signed-off-by: Uli Schlachter <psychon@znc.in>