x11rb: Update default cursor search path
In [1], xcb-util-cursor's default search path was changed. This change
was done to bring it in line with libXcursor's behaviour. In that
library, in [2], ~/.local/share/icons was added to the default search
path.

Additionally, /usr/X11R6/lib/X11/icons was never part of libXcursor's
search path. It only searched there if you set --prefix or --datadir
suitably, but in that case /usr/share/icons would disappear from the
search path. This path was thus removed from xcb-util-cursor in [1].

In this commit, I bring x11rb in sync with xcb-util-cursor again.

[1]: https://gitlab.freedesktop.org/xorg/lib/libxcb-cursor/-/commit/d28a3227eb9e4bdeeb4dd93017c298808bbe62ee
[2]: https://gitlab.freedesktop.org/xorg/lib/libxcursor/-/commit/2263c196cb0d

Signed-off-by: Uli Schlachter <psychon@znc.in>